+++ NEWS.md
@@ -1,3 +1,12 @@
+# Version 0.5.0 [2023-01-30]
+
+### New Features
+
+ * Environments variables to be used during checks can be set in an
+   Renviron file named `revdep/revdepcheck.Renviron`.  This file is
+   loaded by `revdepcheck::check()`.
+
+
 # Version 0.4.0 [2022-11-21]
 
 ### New Features